requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
SQLAlchemy>=2.0
psycopg[binary]
pymongo[srv]
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
